adjust wrapping to remove extra space from paths
It seems that hardwrapping when the newline happens to be inside of an
inline code span is causing an extra space to show up in the rendered
version. For example, I noticed this issue after copy-pasting
`Ghosttty.app/Contents/ Resources/ghostty/themes` from the docs.

There might be other cases in the docs where wrapping is similarly a
problem, I just happened to notice these two.