Add a Link to .deb for Ubuntu
https://github.com/mkasberg/ghostty-ubuntu

Ubuntu is still one of the most popular Linux distributions, and I want
to make it easier for everyone to install Ghostty on Ubuntu. Getting
official Ubuntu packages (perhaps even PPAs) will take some time because
Ubuntu still does not have official zig packages. But we can build a
.deb on GitHub Actions and make it available to download on GitHub.
That's the approach I'm taking in the above repo, which will hopefully
be a good interim solution. (All of the usual caveats about unofficial
packages still apply.)