Fix aria-hidden error
When the fix was applied for screen readers we triggered an error
because the container of the terminal interface had a tab index, which
it really shouldn't since there's nothing to interact with.

I've removed this tab index and also made the quick change to force the
various header bar states to not have a toggle on hover since 1. it's a
bit distracting and 2. sort of unnecessary.

Also it looks like there were some prettier changes that tweaked parts
of the file.