Update terminfo.mdx
This commit adds a note about where the `tic` command writes its output.  I was nervous about access to the system terminfo bits and spent a few minutes reading the terminfo man pages to convince myself that the command wouldn't surprise me.

Thought it might be useful to add a note for the next folks to come along.